# [tpcds](nereids) estimate distribution cost by byte size instead of row count