--- a/fe/fe-core/src/main/java/org/apache/doris/system/SystemInfoService.java
+++ b/fe/fe-core/src/main/java/org/apache/doris/system/SystemInfoService.java
@@ -359,7 +359,7 @@ public class SystemInfoService {
 
     public int getBackendsNumber(boolean needAlive) {
         int beNumber = ConnectContext.get().getSessionVariable().getBeNumberForTest();
-        if (beNumber == -1) {
+        if (beNumber < 0) {
             beNumber = getAllBackendIds(needAlive).size();
         }
         return beNumber;