--- a/regression-test/data/nereids_tpcds_shape_sf100_p0/shape/query31.out
+++ b/regression-test/data/nereids_tpcds_shape_sf100_p0/shape/query31.out
@@ -49,27 +49,27 @@ CteAnchor[cteId= ( CTEId#6=] )
 --------------PhysicalDistribute
 ----------------PhysicalProject
 ------------------hashJoin[INNER_JOIN](ss2.ca_county = ss3.ca_county)
---------------------PhysicalProject
-----------------------filter((ss3.d_year = 2000)(ss3.d_qoy = 3))
-------------------------CteConsumer[cteId= ( CTEId#6=] )
 --------------------PhysicalDistribute
-----------------------hashJoin[INNER_JOIN](ss1.ca_county = ss2.ca_county)(CASE WHEN (web_sales > 0.00) THEN (cast(web_sales as DECIMALV3(38, 8)) / web_sales) ELSE NULL END > CASE WHEN (store_sales > 0.00) THEN (cast(store_sales as DECIMALV3(38, 8)) / store_sales) ELSE NULL END)
+----------------------PhysicalProject
+------------------------filter((ss3.d_year = 2000)(ss3.d_qoy = 3))
+--------------------------CteConsumer[cteId= ( CTEId#6=] )
+--------------------hashJoin[INNER_JOIN](ss1.ca_county = ss2.ca_county)(CASE WHEN (web_sales > 0.00) THEN (cast(web_sales as DECIMALV3(38, 8)) / web_sales) ELSE NULL END > CASE WHEN (store_sales > 0.00) THEN (cast(store_sales as DECIMALV3(38, 8)) / store_sales) ELSE NULL END)
+----------------------PhysicalDistribute
 ------------------------PhysicalProject
 --------------------------filter((ss2.d_year = 2000)(ss2.d_qoy = 2))
 ----------------------------CteConsumer[cteId= ( CTEId#6=] )
+----------------------hashJoin[INNER_JOIN](ss1.ca_county = ws1.ca_county)
 ------------------------PhysicalDistribute
---------------------------hashJoin[INNER_JOIN](ss1.ca_county = ws1.ca_county)
-----------------------------PhysicalDistribute
-------------------------------PhysicalProject
---------------------------------filter((ss1.d_year = 2000)(ss1.d_qoy = 1))
-----------------------------------CteConsumer[cteId= ( CTEId#6=] )
-----------------------------hashJoin[INNER_JOIN](ws1.ca_county = ws2.ca_county)
-------------------------------PhysicalDistribute
---------------------------------PhysicalProject
-----------------------------------filter((ws1.d_year = 2000)(ws1.d_qoy = 1))
-------------------------------------CteConsumer[cteId= ( CTEId#7=] )
-------------------------------PhysicalDistribute
---------------------------------PhysicalProject
-----------------------------------filter((ws2.d_qoy = 2)(ws2.d_year = 2000))
-------------------------------------CteConsumer[cteId= ( CTEId#7=] )
+--------------------------PhysicalProject
+----------------------------filter((ss1.d_year = 2000)(ss1.d_qoy = 1))
+------------------------------CteConsumer[cteId= ( CTEId#6=] )
+------------------------hashJoin[INNER_JOIN](ws1.ca_county = ws2.ca_county)
+--------------------------PhysicalDistribute
+----------------------------PhysicalProject
+------------------------------filter((ws1.d_year = 2000)(ws1.d_qoy = 1))
+--------------------------------CteConsumer[cteId= ( CTEId#7=] )
+--------------------------PhysicalDistribute
+----------------------------PhysicalProject
+------------------------------filter((ws2.d_qoy = 2)(ws2.d_year = 2000))
+--------------------------------CteConsumer[cteId= ( CTEId#7=] )