… regression-test/data/nereids_tpcds_shape_sf100_p0/shape/query44.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/shape/query46.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/shape/query56.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/shape/query59.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/shape/query6.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/shape/query60.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/shape/query65.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/shape/query72.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/shape/query74.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/shape/query81.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/shape/query84.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/shape/query87.out
… regression-test/data/nereids_tpcds_shape_sf100_p0/shape/query91.out
… regression-test/data/nereids_tpch_shape_sf1000_p0/shape/q10.out
… regression-test/data/nereids_tpch_shape_sf1000_p0/shape/q16.out
… regression-test/data/nereids_tpch_shape_sf1000_p0/shape/q2.out
… regression-test/data/nereids_tpch_shape_sf500_p0/shape/q10.out
… regression-test/data/nereids_tpch_shape_sf500_p0/shape/q16.out
… regression-test/data/nereids_tpch_shape_sf500_p0/shape/q2.out