## Proposed changes
this pr impacts tpch q16 Agg strategy, but no performance issue
this pr improves tpcds sf100
before: 
cold 141 sec
hot 133 sec 

after:
code 137 sec
hot 128 sec
Issue Number: close #xxx

<!--Describe your changes.-->
## Further comments

If this is a relatively large or complex change, kick off the discussion at [dev@doris.apache.org](mailto:dev@doris.apache.org) by explaining why you chose the solution you did and what alternatives you considered, etc...